# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kairos API
  description: |
    API for Kairos - Chinese/Japanese immersion learning platform.

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Rate Limiting
    API requests are rate limited. Check response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Window reset time (ISO 8601)

    ## Errors
    All errors return a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message"
      },
      "meta": {
        "requestId": "uuid"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Kairos Support
    email: support@kairos.dev
  license:
    name: Proprietary
servers:
  - url: https://api.kairos.dev/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /cards:
    get:
      tags: [Cards]
      summary: Get all cards
      operationId: getCards
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, nextReview, word]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Cards]
      summary: Create a new card
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Card:
      type: object
      properties:
        id:
          type: string
          format: uuid
        word:
          type: string
        sentence:
          type: string
        pinyin:
          type: string
        definitions:
          type: array
          items:
            type: string
        hskLevel:
          type: integer
          minimum: 1
          maximum: 9
        nextReview:
          type: string
          format: date-time
        interval:
          type: integer
        easeFactor:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateCardRequest:
      type: object
      required: [word, sentence]
      properties:
        word:
          type: string
          minLength: 1
        sentence:
          type: string
          minLength: 1
        pinyin:
          type: string
        definitions:
          type: array
          items:
            type: string
        source:
          type: string
        sourceUrl:
          type: string
          format: uri
    CardResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Card'
    CardsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            cards:
              type: array
              items:
                $ref: '#/components/schemas/Card'
            total:
              type: integer
            hasMore:
              type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          type: object
          properties:
            requestId:
              type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []